requirements.txt
git+https://www.github.com/hbldh/hitherdither
pillow
